requirements.txt
flask==3.1.0
flask-cors==5.0.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0
python-dotenv==1.0.1
requests==2.32.3
pymupdf==1.25.3
python-docx==1.1.2
